+++ index.js
@@ -1,4 +1,16 @@
-import { GenerateGenerator, LfField, LfGroup, LfColumns, LfColumn, LfArray, LfTabs, LfTab, LfSteps, LfStep } from './generator';
+import {
+  GenerateGenerator,
+  LfField,
+  LfGroup,
+  LfColumns,
+  LfColumn,
+  LfArray,
+  LfTabs,
+  LfTab,
+  LfSteps,
+  LfStep,
+  LfObject
+} from './generator';
 
 import { Fields as FieldsReactRsuite5, Forms as FormsReactRsuite5 } from './react-rsuite5/index';
 import { Fields as FieldsReactMaterialUI, Forms as FormsReactMaterialUI } from './react-material-ui/index';
@@ -32,4 +44,4 @@ const FormGenerator = GenerateGenerator({ Fields, Forms });
 export default FormGenerator ;
 export * from './helpers';
 export * from './costants';
-export { Fields, Forms, LfField, LfGroup, LfColumns, LfColumn, LfArray, LfTabs, LfTab, LfSteps, LfStep };
+export { Fields, Forms, LfField, LfGroup, LfColumns, LfColumn, LfArray, LfTabs, LfTab, LfSteps, LfStep, LfObject };